+++ fk/hal/memory.h
@@ -78,13 +78,13 @@ public:
 
     virtual flash_geometry_t geometry() const = 0;
 
-    virtual size_t read(uint32_t address, uint8_t *data, size_t length) = 0;
+    virtual int32_t read(uint32_t address, uint8_t *data, size_t length) = 0;
 
-    virtual size_t write(uint32_t address, uint8_t const *data, size_t length) = 0;
+    virtual int32_t write(uint32_t address, uint8_t const *data, size_t length) = 0;
 
-    virtual size_t erase_block(uint32_t address) = 0;
+    virtual int32_t erase_block(uint32_t address) = 0;
 
-    virtual size_t flush() = 0;
+    virtual int32_t flush() = 0;
 
     bool available() {
         return geometry().total_size > 0;
@@ -106,13 +106,13 @@ public:
 
     flash_geometry_t geometry() const override;
 
-    size_t read(uint32_t address, uint8_t *data, size_t length) override;
+    int32_t read(uint32_t address, uint8_t *data, size_t length) override;
 
-    size_t write(uint32_t address, uint8_t const *data, size_t length) override;
+    int32_t write(uint32_t address, uint8_t const *data, size_t length) override;
 
-    size_t erase_block(uint32_t address) override;
+    int32_t erase_block(uint32_t address) override;
 
-    size_t flush() override;
+    int32_t flush() override;
 
 };
