+++ fk/storage/sequential_memory.h
@@ -12,9 +12,9 @@ public:
     SequentialMemory(DataMemory *memory);
 
 public:
-    size_t read(uint32_t address, uint8_t *data, size_t length) const;
-    size_t write(uint32_t address, uint8_t *data, size_t length);
-    size_t flush();
+    int32_t read(uint32_t address, uint8_t *data, size_t length) const;
+    int32_t write(uint32_t address, uint8_t *data, size_t length);
+    int32_t flush();
 
 };
