+++ fk/storage/statistics_memory.h
@@ -40,13 +40,13 @@ public:
 
     flash_geometry_t geometry() const override;
 
-    size_t read(uint32_t address, uint8_t *data, size_t length) override;
+    int32_t read(uint32_t address, uint8_t *data, size_t length) override;
 
-    size_t write(uint32_t address, const uint8_t *data, size_t length) override;
+    int32_t write(uint32_t address, const uint8_t *data, size_t length) override;
 
-    size_t erase_block(uint32_t address) override;
+    int32_t erase_block(uint32_t address) override;
 
-    size_t flush() override;
+    int32_t flush() override;
 
     MemoryStatistics &statistics();
